requirements.txt
torch
torchvision
torchaudio
mlflow
fastapi
uvicorn[standard]
pillow
pyyaml
pytest
dvc
